correction display bug

+++ interface.py
@@ -23,12 +23,11 @@ class Interface:
         self._current = 33
         self._root.title('Seselab')
         self._dm = 1
-        self.open_reglist()
+        self.creat_reglist()
         self.font = tkFont.nametofont("TkFixedFont")
         self.font.configure(size=14)
         self.which_file()
-        self._inject_fault = Label(self._root, text = 'Inject fault:', bg = None, width = 10, font = ('helvetic', 11, 'bold'))
-        self._inject_fault.grid(column = 1, row = 10, columnspan = 8, sticky = 'w', padx = 8)
+
         self._step = self.creat_button(self._root, 'Step', self.event_step, 'disabled', 8, 1)
         self._run = self.creat_button(self._root, 'Run', self.event_run, 'disabled', 8, 2)
         self._run_slow = self.creat_button(self._root, 'Run slowly', self.event_run_slow, 'disabled', 8, 3)
@@ -36,17 +35,20 @@ class Interface:
         self._skip = self.creat_button(self._root, 'Skip', self.event_skip, 'disabled', 8, 5)
         self._to_rand = self.creat_button(self._root, 'Random', self.event_set_rand, 'disabled', 2, 10)
         self._to_zero = self.creat_button(self._root, 'Zero', self.event_set_zero, 'disabled', 3, 10)
-        self._reset = self.creat_button(self._root, 'Reset', self.load, 'disabled', 5, 0)
+        self._reset = self.creat_button(self._root, 'Reset', self.event_reset, 'disabled', 5, 0)
         self._quit= self.creat_button(self._root, 'Close', self._root.quit, 'normal', 8, 12)
         self._pick_file = self.creat_button(self._root, 'Pick file', self.pick_file, 'normal', 3, 0)
-        self._load = self.creat_button(self._root, 'Load', self.load, 'normal', 4, 0)
+        self._load = self.creat_button(self._root, 'Load', self.event_load, 'normal', 4, 0)
 
+        self._inject_fault = Label(self._root, text = 'Inject fault:', bg = None, width = 10, font = ('helvetic', 11, 'bold'))
+        self._inject_fault.grid(column = 1, row = 10, columnspan = 8, sticky = 'w', padx = 8)
+        
         self._canvas = Canvas(self._root, borderwidth = 0, height = 150, background="#ffffff")
         self._frame = Frame(self._canvas, background = "#ffffff")
         self._vsb = Scrollbar(self._root, orient = "vertical", command = self._canvas.yview)
         self._canvas.configure(yscrollcommand = self._vsb.set)
         self._vsb.grid(rowspan = 5, column = 7, row = 1, sticky = 'nse')
-        self._canvas.grid(rowspan = 5, columnspan = 7,column = 1, row = 1, sticky = 'nesw')
+        self._canvas.grid(rowspan = 5, columnspan = 7,column = 1, row = 1, sticky = 'nesw', pady = 5, padx = 5)
         self._canvas.create_window((4,4), window=self._frame, anchor="ne")
         self._frame.bind("<Configure>", lambda event, canvas = self._canvas: self.on_frame_configure(self._canvas))
         self._op = Label(self._root, text = '', anchor = 'w', bg = 'white', width = 85, height = 5, font = self.font, relief = SUNKEN)
@@ -62,7 +64,8 @@ class Interface:
                       bg = 'white',
                       font = self.font)
             l.grid(row = row, 
-                   column = 1)
+                   column = 1,
+                   pady = 2)
             self._istr_lst.append(l)
 
     def on_frame_configure (self, canvas):
@@ -77,45 +80,69 @@ class Interface:
         button.grid(column = c, 
             row = r, 
             sticky = 'we', 
-            pady = 0, 
-            padx = 2)
+            padx = 5)
         return button
 
     def pick_file (self):
         filename = filedialog.askopenfilename(initialdir = os.getcwd(), 
             title = "Select file", 
             filetypes = (("asm files", "*.asm"), ("all files", "*.*")))
-        self._file.delete(0, END)
-        self._file.insert(0, filename)
+        # if filename != '':
+        if len(filename) > 0:
+            self._file.delete(0, END)
+            self._file.insert(0, filename)
 
     def which_file (self):
         value = StringVar() 
         try:
             value.set(sys.argv[1])
-            self._file = Entry(self._root, text = value, width = 22, font = ('helvetic', 12))
-            self._file.grid(column = 1, row = 0, sticky = 'w', padx = 7, pady = 5, columnspan = 3)
+            self._file = Entry(self._root, 
+            	text = value, 
+            	width = 22, 
+            	font = ('helvetic', 12))
+            self._file.grid(column = 1, 
+            	row = 0, 
+            	sticky = 'w', 
+            	padx = 7, 
+            	pady = 5, 
+            	columnspan = 3)
         except:
             value.set('Choose a file ...')
-            self._file = Entry(self._root, text=value, width = 22, font = ('helvetic', 12))
-            self._file.grid(column = 1, row = 0, sticky = 'w', padx = 7, pady = 5, columnspan = 3)
+            self._file = Entry(self._root, 
+            	text=value, 
+            	width = 22, 
+            	font = ('helvetic', 12))
+            self._file.grid(column = 1, 
+            	row = 0, 
+            	sticky = 'w', 
+            	padx = 7, 
+            	pady = 5, 
+            	columnspan = 3)
 
     def event_reset (self):
         self._i = 0
+        self._sortie = ''
         self.cpu = CPU(1048576, 32, self._code, '/dev/null')
         self.cpu._ip += 1
         self._infos = Instr().stack_infos(self._code)
+        self._canvas.yview_scroll(-2*len(self._infos),'units')
         self.populate()
         self.update_display()
+        self.button_state('normal')
+        self._pause.config(state = 'normal')
+        self._reset.config(state = 'normal')
         self.update_reglist()
         self.flag = 0
 
-    def load (self):
+
+    def event_load (self):
         self._i = 0
         self._sortie = ''
         self._code = Compiler().compile(self._file.get())
         self.cpu = CPU(1048576, 32, self._code, '/dev/null')
         self.cpu._ip += 1
         self._infos = Instr().stack_infos(self._code)
+        self._canvas.yview_scroll(-2*len(self._infos),'units')
         self.populate()
         self.update_display()
         self.button_state('normal')
@@ -142,7 +169,7 @@ class Interface:
             self._to_rand.config(state = 'normal')
             self._to_zero.config(state = 'normal')
 
-    def open_reglist (self):
+    def creat_reglist (self):
         self._reg_lst = []
         i = 0
         for j in range(4):
@@ -167,15 +194,15 @@ class Interface:
     def update_display (self):
         for i in range (32):
             self._reg_lst[i].config(text='r'+str(i)+': '+str(self.cpu._reg[i]))
-            self._canvas.yview_scroll(-2*i,'units')
-        if self._i < 3:
+        self._canvas.yview_scroll(-2*len(self._infos),'units')
+        if self._i < 2:
             self._istr_lst[self._i].config(bg = 'white')
             self._istr_lst[self._i+1].config(bg = 'yellow')
-        elif self._i >= len(self._infos)-4:
+        elif self._i >= len(self._infos)-2:
             self._istr_lst[self._i].config(bg = 'white')
             self._istr_lst[self._i+1].config(bg = 'yellow')
         else:
-            self._istr_lst[self._i-3].destroy()
+            self._canvas.yview_scroll(2*self._i-2,'units')
             self._istr_lst[self._i].config(bg = 'white')
             self._istr_lst[self._i+1].config(bg = 'yellow')
         self._i += 1
@@ -207,6 +234,9 @@ class Interface:
                 self.update_display()
                 self._root.after(500,self.event_tempo)
                 pass
+            else:
+                self._load.config(state = 'normal')
+                self._reset.config(state = 'normal')
 
     def event_run (self):
         self._speed = 0
